+++ tests/__init__.py
@@ -1,0 +1,4 @@
+"""
+AM-Corp Test Suite
+"""
+